refactor: add task utils
